Create app3.py
this is able to fetch the reddit URL's, but the issue is that when I click the regenerate button, the process starts all over instead of just picking from the already generated list. Also need to figure out how to relink the google searches. But saving seems to work fine.